Move shared into client for create-react-app
