+++ client/src/IntroductionPhase.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import './IntroductionPhase.css';
-import {Interactions} from '../../shared/data.js';
+import {Interactions} from './shared/data.js';
 import Delay from 'react-delay';
 import Swipeable from './components/Swipeable.js';
