+++ client/src/loaders/loadDataForCohort.js
@@ -3,7 +3,7 @@ import __flatten from 'lodash/flatten';
 import parseCsvSync from 'csv-parse/lib/sync';
 import profileTemplatesFile from '../files/profileTemplates.csv';
 import sortedVariantsFile from '../files/sortedVariants.csv';
-import {hashCode} from '../../../shared/data.js';
+import {hashCode} from '../shared/data.js';
 import {createProfiles} from './createProfiles.js';
